ACPI: PCI: hotplug: Avoid setting is_hotplug_bridge for PCIe Upstream Ports
It is reported that on some systems pciehp binds to an Upstream Port and
attempts to operate it which causes devices below the Port to disappear
from the bus.

This happens because acpiphp sets is_hotplug_bridge for that Port (after
receiving a Device Check notification on it from the platform firmware
via ACPI) during the enumeration of PCI devices and so when
get_port_device_capability() runs, it sees that is_hotplug_bridge is
set and adds PCIE_PORT_SERVICE_HP to Port services (which allows pciehp
to bind to the Port in question).

Even though this particular problem can be addressed by making the
portdrv_core checks more robust, it also causes power management to
work differently on the affected systems which generally is not
desirable (PCIe Ports with is_hotplug_bridge set have to pass
additional tests to be allowed to go into the D3hot/cold power
states which affects runtime PM of devices below these Ports).

For this reason, amend check_hotplug_bridge() with a PCIe type check
to prevent it from setting is_hotplug_bridge for Upstream Ports.

Reported-by: Rodrigo Vivi <rodrigo.vivi@intel.com>
Suggested-by: Lukas Wunner <lukas@wunner.de>
Signed-off-by: Rafael J. Wysocki <rafael.j.wysocki@intel.com>