drm/xe: Include documentation from source in rendered output
Include all documentation in the rendered output, except the ones
with GuC ABI documentation (drivers/gpu/drm/xe/abi/). These probably
need to be a added in a single page, but need to be revisited if they
are required.

Also follow Documentation/doc-guide/kernel-doc.rst so the titles are in the
recommended order.

Signed-off-by: Lucas De Marchi <lucas.demarchi@intel.com>