drm/xe: Fix module unload, finally
intel_fbdev needs to unpin the BO before unmapping. Small easy fix.

The display implementation links to a bunch of components and other
things that need to be torn down first, so our normal teardown
won't happen until we first break the dependency chain.

We need to nuke at least fbdev, but I think it's best to kill other
components too, just to be thorough.

Also add a lastclose handler to restore fbdev.

Signed-off-by: Maarten Lankhorst <maarten.lankhorst@linux.intel.com>