ALSA: hda/hdmi: fix setting stream id when in D3
When the new style KAE keep-alive implementation is used, the clocks are
maintained when codec is in D3. The generic code in
hda_cleanup_all_streams() can however interfere with generation of audio
samples in this mode, by setting the stream and channel ids to zero.

To get full benefit of the keepalive, set the new
no_stream_clean_at_suspend quirk bit on affected Intel hardware. When
this bit is set, stream cleanup is skipped in hda_call_codec_suspend().

Special handling is needed for the case when system goes to suspend. The
stream id programming can be lost in this case. This will also cause
codec->cvt_setups to be out of sync. Handle this by reprogramming silent
streaming upon resume.

Fixes: 15175a4f2bbb ("ALSA: hda/hdmi: add keep-alive support for ADL-P and DG2")
Signed-off-by: Kai Vehmanen <kai.vehmanen@linux.intel.com>