drm/xe: Don't evict same vm bos when validating for exec or rebind worker
When validating bos for exec and rebind worker we allowed evicting bos
from the same VM in the process. These bos were then added to the
validation list and that could in theory cause an infinite validation
loop.

In fault mode, however, we still want to allow evicting bos using the
same VM in the fault handler.

So fix this by adding an additional argument to xe_bo_validate() and
disable evictions from the same vm during exec- and rebind worker
validation.

Signed-off-by: Thomas Hellström <thomas.hellstrom@linux.intel.com>