drm/xe: Use a list for the vm's external objects
The way we handle the vm's external object array is of O(n) complexity
both when it comes to addition and removal of a single object, so use a
list of external bo vmas instead, which is O(1).

This also paves the way for introducing VM fences, which can, to some
extent at least, remove the need for traversing the external object list
on each exec, but that's for later.

Signed-off-by: Thomas Hellström <thomas.hellstrom@linux.intel.com>