drm/xe: improve error capture by dumping vma
Still under XE_SIMPLE_ERROR_CAPTURE, but that prevents
heavy instrumentation of user-side code. By sharing the
addresses, we can check the content from the user side
using a simple application (from the table of pages to
the contents of buffers). The error dump fails if the
vm lock cannot be acquired.

Co-developed-by: Philippe Lecluse <philippe.lecluse@intel.com>
Signed-off-by: Philippe Lecluse <philippe.lecluse@intel.com>
Signed-off-by: Francois Dugast <francois.dugast@intel.com>