drm/xe: Remove Author from xe_res_cursor.h header.
We need to keep the license and the copyright notice, but the
Author name in there we should remove since he was not the
one writing this file specifically. And also because we are not
using this Author in the files any longer.

I was planning to remove this when squashing the intial patch,
but I'm afraid that I'll end up forgetting to do it, so let's
do it now.

Signed-off-by: Rodrigo Vivi <rodrigo.vivi@intel.com>