drm/xe: Introduce a new DRM driver for Intel GPUs
Xe, is a new driver for Intel GPUs that supports both integrated and
discrete platforms starting with Tiger Lake (first Intel Xe Architecture).

The code is at a stage where it is already functional and has experimental
support for multiple platforms starting from Tiger Lake, with initial
support implemented in Mesa (for Iris and Anv, our OpenGL and Vulkan
drivers), as well as in NEO (for OpenCL and Level0).

The new Xe driver leverages a lot from i915.

As for display, the intent is to share the display code with the i915
driver so that there is maximum reuse there.

This initial work is a collaboration of many people and unfortunately
the big squashed patch won't fully honor the proper credits. But let's
get some git quick stats so we can at least try to preserve some of the
credits:

Co-developed-by: Matthew Brost <matthew.brost@intel.com>
[commits: 1105 / lines changed: 101133]
Co-developed-by: Matthew Auld <matthew.auld@intel.com>
[commits: 1 / lines changed: 2]
Co-developed-by: Matt Roper <matthew.d.roper@intel.com>
[commits: 6 / lines changed: 936]
Co-developed-by: Thomas Hellström <thomas.hellstrom@linux.intel.com>
[commits: 105 / lines changed: 11705]
Co-developed-by: Francois Dugast <francois.dugast@intel.com>
[commits: 11 / lines changed: 308]
Co-developed-by: Lucas De Marchi <lucas.demarchi@intel.com>
[commits: 142 / lines changed: 9413]
Co-developed-by: Maarten Lankhorst <maarten.lankhorst@linux.intel.com>
[commits: 143  / lines changed: 20119]
Co-developed-by: Philippe Lecluse <philippe.lecluse@intel.com>
[commits: 156 / lines changed: 6930]
Co-developed-by: Nirmoy Das <nirmoy.das@intel.com>
[commits: 3 / lines changed: 72]
Co-developed-by: Jani Nikula <jani.nikula@intel.com>
[commits: 6 / lines changed: 78]
Co-developed-by: José Roberto de Souza <jose.souza@intel.com>
[commits: 36 / lines changed: 836]
Co-developed-by: Rodrigo Vivi <rodrigo.vivi@intel.com>
[commits: 58 / lines changed: 7618]
Co-developed-by: Dave Airlie <airlied@redhat.com>
[commits: 66 / lines changed: 2742]
Co-developed-by: Faith Ekstrand <faith.ekstrand@collabora.com>
[commits: 390 / lines changed: 108951]
Co-developed-by: Daniel Vetter <daniel.vetter@ffwll.ch>
Signed-off-by: Rodrigo Vivi <rodrigo.vivi@intel.com>
Signed-off-by: Matthew Brost <matthew.brost@intel.com>
Signed-off-by: Rodrigo Vivi <rodrigo.vivi@intel.com>