drm/xe/bo: Scrub preempt fences in the TTM release_notify callback.
We don't want bo destruction to unnecessarily trigger compute engine
pausing. Since the unbind fence is already attached to the resv,
just scrub the preempt fences from the resv once we're sure it's been
individualized.

Locking in the destructor is a little tricky: Even blocking locks should
always succeed here, but we might get false lockdep splats because we
might be holding the vm's resv while putting the bo.

Just use the same way TTM does its locking of objects on the delayed
destoy lists; trylocking under the lru lock. We should remember to
fix this up to be less fragile to changes in TTM.

Signed-off-by: Thomas Hellström <thomas.hellstrom@linux.intel.com>