drm/xe: Hold vm->lock across retry in preempt rebind reworker
If we drop lock in the retry loop it is possible to create a new preempt
fence with a resume state and an exec to come in which submits a job to
the newly created engine while the rebind worker is still running. This
is not allowed, the rebind worker must complete before any preempt fence
can be in the resume state.  To avoid this hold vm->lock in the retry to
block the creation of new preempt fences while it is running.

Because we must hold the lock across the entire rebind worker, scheme of
down grading the write lock to read lock falls apart. For simplicity
just take the vm->lock in write mode in the rebind worker.

Signed-off-by: Matthew Brost <matthew.brost@intel.com>