drm/xe: Get DG2 back to life.
After commit "drm/xe: Update VRAM manager to use buddy allocator"
both S4 and D3cold got broken.

Actually even with d3hot on DG2, if we get the pci_disable_device
and then we try to run xe_bo_restore_kernel it won't complete
and return an error, breaking any attempt of resume.

This affects the D3 flow and the S4 flow as well. For some
reason it is not impacting S3 and S2idle. More debug is needed
but for now, let's disable the d3cold_allow so we get DG2 back
to an usable point if Xe is compiled without Display.

With display the runtime_pm cannot suspend hence the issue
was not seen.

Signed-off-by: Rodrigo Vivi <rodrigo.vivi@intel.com>