drm/xe/reg_sr: Stop using i915_reg_t
i915_reg_t is not representative of all the registers anymore.  i915
also has i915_mcr_reg_t for some of them. For the rtp and reg_sr infra
in xe, defining a separate type for them is not beneficial.

For reg_sr, the use of i915_reg_t is mainly to create an index for the
xarray. Any additional data can be stored in the entry.  The index can
be just a plain u32.

Signed-off-by: Lucas De Marchi <lucas.demarchi@intel.com>
Reviewed-by: Rodrigo Vivi <rodrigo.vivi@intel.com>