drm/xe/rtp: Stop using i915 reg types
Add xe_rtp_reg_t and start using it for the RTP tables. Since the
knowledge of MCR register is in the type, provide alternative macros to
be used in place of _MMIO()/MCR_REG() and just use plain u32.

These redefintions need to be done before defining the table and failure
to do so will just break compilation as the previous types (i915_reg_t
and i915_mcr_reg_t) are not compatiable with the new one.

With the new macros the MCR annotation goes to a separate field which
will be used later to implement read/write of MCR registers in reg_sr.

Signed-off-by: Lucas De Marchi <lucas.demarchi@intel.com>
Reviewed-by: Rodrigo Vivi <rodrigo.vivi@intel.com>