drm/xe: Move whitelist application to xe_reg_sr
The application of a mmio write should be in xe_reg_sr.c, i.e. "register
save/restore" rather than in the xe_wa.c file. The later is dedicated to
contain the RTP tables with WAs.

Besides making it more correct from the organization point of view, this
is important as there will be need to overwrite _MMIO() and MCR_REG()
macros in order to support mcr-style register. Since the application of
the register uses those macros, it can't be in that file.

Signed-off-by: Lucas De Marchi <lucas.demarchi@intel.com>
Reviewed-by: José Roberto de Souza <jose.souza@intel.com>