drm/xe: Rework initial plane readout GGTT mapping
We cannot re-use the initial position of the GGTT, at least not on TGL,
as it's put immediately into the beginning of GGTT, which is reserved
for WOPCM.

Since on DGFX there is no 1:1 mapping between VRAM and GGTT, there's
also the potential of a similar bug in DG2 where it attempts to scanout
scratch pages.

To prevent all of this, we allocate a second GGTT entry at a
higher position, and immediately force an async flip to the new
position.

As a small optimization, we also force pin the initial scanout GGTT
on all platforms, and make the pin to GGTT a NOOP.

Signed-off-by: Maarten Lankhorst <maarten.lankhorst@linux.intel.com>
Reviewed-by: Matthew Auld <matthew.auld@intel.com>