drm/xe/display: Rename and clarify xe_display_enable()
Rename xe_display_enable() so it doesn't appear that the IP is being
enabled at this point: it's very early in the initialization phase and
this is actually setting up the SW part in the DRM driver to be able do
drive display. While at it, add the kernel doc.

Signed-off-by: Lucas De Marchi <lucas.demarchi@intel.com>
Reviewed-by: Maarten Lankhorst <maarten.lankhorst@linux.intel.com>
Link: https://lore.kernel.org/r/20230302013411.3262608-4-lucas.demarchi@intel.com