drm/xe/display: Move display sw init to xe_display.c
Instead of having all the fields initialized in xe_device_create(), move
the display part to xe_display_create(). Right now the name "_create()"
is not so good as it's basically initializing the fields from xe_device.
There is still work needed to move fields to the display substruct so it
can be an opaque struct.  However, there are already several _init_*()
functions, so use the create() suffix in the hope the struct will be
made opaque soon.

Signed-off-by: Lucas De Marchi <lucas.demarchi@intel.com>
Reviewed-by: Maarten Lankhorst <maarten.lankhorst@linux.intel.com>
Link: https://lore.kernel.org/r/20230302013411.3262608-5-lucas.demarchi@intel.com