drm/xe/display: Move device info initialization to display
Instead of initializing the display info together with the platform in
xe_pci.c, let the display part initialize it. The small disadvantage is
that the description struct is not automatically mapped from the PCI ID,
but just doing a switch on the platform, that should already be set
during display initialization should be ok.

This allows to encapsulate the display details inside the display
compilation units.

Also use __diag_push() / __diag_pop() like in xe_step to handle the few
places where -Woverride-init should be disabled.

Signed-off-by: Lucas De Marchi <lucas.demarchi@intel.com>
Reviewed-by: Maarten Lankhorst <maarten.lankhorst@linux.intel.com>
Link: https://lore.kernel.org/r/20230302013411.3262608-6-lucas.demarchi@intel.com
Signed-off-by: Rodrigo Vivi <rodrigo.vivi@intel.com>