drm/xe: Add VM snapshot to xe_devcoredump.
With this patch, we now have some parity between xe_devcoredump
and the simple_error_capture. The only difference is that
xe_devcoredump will only stash the 'first' hang, which is the one
that we care most and should analyze first, while
simple_error_capture will dump them all the kernel log.

But this is just a start point to start building a useful and
organized crash dump, using standard infrastructure. Later this
will be changed to have output that can be parsed by tools and
used for error replay.

Also, it is important to highlight that the goal is not to replace
the simple_error_capture which is still useful for some cases.
But simple_error_capture should be protected under DEBUG and
EXPERT flags, while the devcoredump has its own production config
and will be useful for bug reporting and for error replay.

v: Removed extra new line (Matthew)

Signed-off-by: Rodrigo Vivi <rodrigo.vivi@intel.com>
Reviewed-by: Matthew Brost <matthew.brost@intel.com>