drm/xe/rtp: Improve magic macros for RTP tables
The macros for writing the rtp tables have some downsides:

	1) They are fragile, and when making a typo it's hard to know
	   where it's failing for someone not used to them
	2) They polute the global scope. The CONCAT/COUNT_ARGS/etc only
	   work because they are defined exactly the same as the ones in
	   kernel.h.
	3) They are not compatible with rules/actions with registers
	   passing additional options

This new implementation fixes (2) and (3).

For (2), prefix everything with "_XE" or "XE_RTP", so there is less chance of
clashes. Also, the maximum number of rules and actions is 4, so there is
no need to make the helper macros go until 12: reduce that to 5, so if
more arguments is passed, the compiler gives an error that there is not
definition with _5 suffix.

For (3), change the implementation to concentrate on "pasting a prefix
to each argument" rather than the more general "call any macro for each
argument". Hopefully this will avoid trying to extending this infra and
making it more complex. The new changes to how the arguments are
constructed using tuples will allow extending the interface where it's
needed: passing additional register fields to allow creating our own
struct xe_reg.

v2:
  - Prefer XE_RTP_ prefix for helpers use in the rest of rtp header
    (Michal Wajdeczko)
  - Remove unnecessary indirection to paste to each arg

Reviewed-by: Matt Roper <matthew.d.roper@intel.com> # v1
Link: https://lore.kernel.org/r/20230427223256.1432787-6-lucas.demarchi@intel.com
Signed-off-by: Lucas De Marchi <lucas.demarchi@intel.com>