drm/xe/display: Use xe_display_driver prefix
Use a separate xe_display_driver prefix for functions dealing with the
DRM driver struct or the early probe. This splits
xe_display_set_driver_hooks() in 2: one to allow the driver to defer the
probe if the display needs that and the other to set the additional
hooks in the driver struct.  Although the hooks are set by every device
binding to the module, they only depend on the module param
enable_display, so that is safe.

Since unset_driver_hooks() is not really tweaking the driver, but rather
the driver_features saved in the device, rename it to
unset_display_features() to clarify that.

Reviewed-by: Matt Atwood <matthew.s.atwood@intel.com>
Link: https://lore.kernel.org/r/20230510195424.3045127-5-lucas.demarchi@intel.com
Signed-off-by: Lucas De Marchi <lucas.demarchi@intel.com>