devcoredump: Add dev_coredump_remove function.
Add a functionality where the failing device can remove the
dev_coredump uppon un-probe.

Since devcoredump is getting the reference for the failing
device, the pci module cannot be removed before the 5 minute
timeout expires, or before the user echo 1 > devcd<n>/data.

On top of that, for PCI devices, the unbind of the device will
call the pci .remove void function, that cannot fail. At that
time, our device is pretty much gone, but the read and free
functions are alive and they can get some NULL dereferences
or use after free.

Signed-off-by: Rodrigo Vivi <rodrigo.vivi@kernel.org>