fixup! drm/xe: Introduce a new DRM driver for Intel GPUs
drm/xe: Fix unlocked access of the vma::rebind_link

the vma::rebind_link is protected by the vm's resv, but we were
modifying it without. Fix this by use the vma::userptr_link instead
for the tmp_evict list. The vma::userptr_link is protected by the
vm lock.

Cc: Oded Gabbay <ogabbay@kernel.org>
Signed-off-by: Thomas Hellström <thomas.hellstrom@linux.intel.com>
Reviewed-by: Matthew Brost <matthew.brost@intel.com>