drm/xe/display: Append dma_buf fences to framebuffer fences
This is required to comply with KMS page flip requirements, and it is
the only way to synchronize framebuffer scanouts in Xe KMD.

UMDs are expected to attach a syncobj that will be signaled at the end
of the GPU job that is rendering in the framebuffer.
This way, the compositor can send jobs to the GPU to render the
framebuffer and without having to wait for it to complete, queue the
page flip in KMS.

This fixes display glitches.

Link: https://gitlab.freedesktop.org/mesa/mesa/-/issues/8923
Reviewed-by: Uma Shankar <uma.shankar@intel.com>
Signed-off-by: José Roberto de Souza <jose.souza@intel.com>