drm/xe: Fix and clarify PAT Index selection.
PAT selection here in the Xe driver had inherited a lot
of history from i915 and ended up with some strange cases
like doing logical 'or' with 0, or even unnecessarily selecting
pat_index = 7, instead of simply opting for the 0 that is the
right one for Xe platforms.

XXX: More work needs to be done in MTL for the right selection,
starting from the fact that pat_index = 2 is now the uncached
and pat_index = 3 is WB with some kind of coherence.

Signed-off-by: Rodrigo Vivi <rodrigo.vivi@intel.com>