ata: libata-scsi: link ata port and scsi device
There is no direct ancestry between an ata_device and its scsi device,
which prevents the power management code from correctly ordering suspend
and resume operations, which requires additional code to be handled
correctly. Create such ancestry to allow simplifying the code.

The parent-child (supplier-consumer) relationship is established between
the ata_port (parent) and the scsi device (child) with
device_add_link(). The parent used is not the ata_device as the PM
operations are defined per port and devices status controlled from these
port operations.

The device link is established with the new function
ata_scsi_dev_alloc(). This function is used to define the ->slave_alloc
callback of the scsi host template of all drivers.

Signed-off-by: Damien Le Moal <dlemoal@kernel.org>