drm/xe/uapi: Rename couple exec_queue items
'Placement' is no used in many terms around the memory_region selection
where the BO or the page table will live. Also, the job itself deserves
a word of more action since it is dispatched to the engine.

'width' is so generic and in graphics world can mean many other different
things. Let's be more specific here on the intent of that.

Signed-off-by: Rodrigo Vivi <rodrigo.vivi@intel.com>