squash! drm/xe/uapi: Rename couple exec_queue items
(squash instead of the fixup so the commit message can be updated)

(new full commit message with new subject)

drm/xe/uapi: Exec queue documentation and variable renaming

Rename 'placement' to num_eng_per_bb and 'width' to num_bb_per_exec, and
give a graphical documentation to it.

Let's make it obvious and straight forward. Not only because it is important
to have variable names that are clear and descriptive, but also because
'placement' is now used in many terms around the memory_region selection
where the BO or the page table will live and 'width' is so generic and with
so many other common meaning in the graphics world.

Signed-off-by: Rodrigo Vivi <rodrigo.vivi@intel.com>