fixup! drm/xe: Make DRM_XE_DEVICE_QUERY_ENGINES future proof
Add more reserved space, since we already know in advance that we
will need to at least show capabilities, because not every engine
even within the same class is able to perform exactly the same actions.

We might also need flags and/or properties. So, let's give us a
bit more room to be really future proof.

Signed-off-by: Rodrigo Vivi <rodrigo.vivi@intel.com>