drm/xe: Runtime PM wake on every debugfs call
Let's ensure our PCI device is awaken on every debugfs call.
Let's increase the runtime_pm protection and start moving
that to the outer bounds.

For now, we are not removing the xe_device_mem_access wrappers
becase the mem_access assert is still checking for its counter
and we don't want to change that behavior just yet. A complete
refactor of that assert and the killing of the mem_access_{get, put}
wrappers are coming soon. Right after we ensure all driver
outer entries are protected with direct pm_runtime references.

Signed-off-by: Rodrigo Vivi <rodrigo.vivi@intel.com>