drm/xe: Runtime PM wake on memory mapped cases for dgfx
Let's ensure our PCI device is awaken on every memory mapped
access on dgfx.
Let's increase the runtime_pm protection and start moving
that to the outer bounds.

No local memory access is allowed when the PCI device is on D3hot.
As stated in the comments, in real world cases we could eventually
rely on display or compute VMs instead of blocking at this mmapped
BO level. However our test infrastructure is not prepared to handle
that, so we are going over protective on discrete, even if this
means less power savings for now.

For now, we are not removing the xe_device_mem_access wrappers
becase the mem_access assert is still checking for its counter
and we don't want to change that behavior just yet. A complete
refactor of that assert and the killing of the mem_access_{get, put}
wrappers are coming soon. Right after we ensure all driver
outer entries are protected with direct pm_runtime references.

Signed-off-by: Rodrigo Vivi <rodrigo.vivi@intel.com>