fixup! drm/xe: Memory allocations are tile-based, not GT-based
Fix a missed conversion of GT to tile in vma_is_valid().
Best is to update the function to take a tile argument.

This bug causes page faults to not complete successfully from media
engines and when the PTE had been invalidated prior.
(Caught by igt@xe_exec_fault_mode and INVALIDATE tests.)

Signed-off-by: Brian Welty <brian.welty@intel.com>
Reviewed-by: Matthew Brost <matthew.brost@intel.com>
Reviewed-by: Matt Roper <matthew.d.roper@intel.com>